add removeUnusedNames staticmethod to name table
